Near complete GTK 3.6 support. Removed Unico engine dependency, using built-in engine.
There are still some lingering issues, namely checks and bullets not showing up in the checkbox and radio boxes. This is likely due to the removal of unico which was capable of drawing checks and bullets itself. I may need to provide my own images for such things. In the meantime, a check or a radio button can still be destinguished when it is active or not because of the color difference.

Signed-off-by: Max Fierke <max@maxfierke.com>